Rename file to match param-case style
